Fix ArrayQ for SparseArray
