+++ mathics/builtin/testing_expressions/list_oriented.py
@@ -7,10 +7,10 @@ from mathics.core.builtin import Builtin, Test
 from mathics.core.evaluation import Evaluation
 from mathics.core.exceptions import InvalidLevelspecError
 from mathics.core.expression import Expression
-from mathics.core.rules import Pattern
 from mathics.core.symbols import Atom, SymbolFalse, SymbolTrue
-from mathics.core.systemsymbols import SymbolSubsetQ
+from mathics.core.systemsymbols import SymbolSparseArray, SymbolSubsetQ
 from mathics.eval.parts import python_levelspec
+from mathics.eval.testing_expressions import check_ArrayQ, check_SparseArrayQ
 
 
 class ArrayQ(Builtin):
@@ -39,6 +39,14 @@ class ArrayQ(Builtin):
      = False
     >> ArrayQ[{{a, b}, {c, d}}, 2, SymbolQ]
      = True
+    >> ArrayQ[SparseArray[{{1, 2} -> a, {2, 1} -> b}]]
+     = True
+    >> ArrayQ[SparseArray[{{1, 2} -> a, {2, 1} -> b}], 1]
+     = False
+    >> ArrayQ[SparseArray[{{1, 2} -> a, {2, 1} -> b}], 2, SymbolQ]
+     = False
+    >> ArrayQ[SparseArray[{{1, 1} -> a, {1, 2} -> b}], 2, SymbolQ]
+     = True
     """
 
     rules = {
@@ -51,37 +59,10 @@ class ArrayQ(Builtin):
     def eval(self, expr, pattern, test, evaluation: Evaluation):
         "ArrayQ[expr_, pattern_, test_]"
 
-        pattern = Pattern.create(pattern)
+        if not isinstance(expr, Atom) and expr.head.sameQ(SymbolSparseArray):
+            return check_SparseArrayQ(expr, pattern, test, evaluation)
 
-        dims = [len(expr.get_elements())]  # to ensure an atom is not an array
-
-        def check(level, expr):
-            if not expr.has_form("List", None):
-                test_expr = Expression(test, expr)
-                if test_expr.evaluate(evaluation) != SymbolTrue:
-                    return False
-                level_dim = None
-            else:
-                level_dim = len(expr.elements)
-
-            if len(dims) > level:
-                if dims[level] != level_dim:
-                    return False
-            else:
-                dims.append(level_dim)
-            if level_dim is not None:
-                for element in expr.elements:
-                    if not check(level + 1, element):
-                        return False
-            return True
-
-        if not check(0, expr):
-            return SymbolFalse
-
-        depth = len(dims) - 1  # None doesn't count
-        if not pattern.does_match(Integer(depth), evaluation):
-            return SymbolFalse
-        return SymbolTrue
+        return check_ArrayQ(expr, pattern, test, evaluation)
 
 
 class DisjointQ(Test):
